+++ packages/utilities/react-rsc-utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./children";
 export * from "./filter-dom-props";
+export * from "./dom-props";