+++ packages/components/switch/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nextui-org/switch
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @nextui-org/react-utils@2.0.4
+
 ## 2.0.5
 
 ### Patch Changes