+++ packages/components/tooltip/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nextui-org/tooltip
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @nextui-org/react-utils@2.0.4
+
 ## 2.0.6
 
 ### Patch Changes